+++ CHANGELOG.md
@@ -35,7 +35,7 @@ Added / Changed / Deprecated / Fixed / Removed / Security
 #### org.ojalgo.matrix
 
 - `MatrixStore` now implements `Structure2D.Logical` directly. No need to call `logical()` to get a `LogicalBuilder`.
-- The `LogicalBuilder`:s of the various `BasicMatrix` subclasses now also implement `Operate2D`.
+- `BasicMatrix` now implements `Structure2D.Logical` as well as `Operate2D` directly. No need to call `logical()` to get a `LogicalBuilder`.
 - A lot of refactoring among the package private code.
 
 #### org.ojalgo.scalar
@@ -54,6 +54,7 @@ Added / Changed / Deprecated / Fixed / Removed / Security
 
 - The `MatrixStore.Factory` interface has been removed. Corresponding functionality have instead been added to `PhysicalStore.Factory`. This also mean that the various static factory instances `MatrixStore` have been removed. Instead use the instances available in each of the `PhysicalStore` instances.
 - The `MatrixStore.LogicalBuilder` class has been removed. Instead `MatrixStore` now implements `Structure2D.Logical` directly. No need to call `logical()` to get a `LogicalBuilder`.
+- The `BasicMatrix.LogicalBuilder` class has been removed...
 
 #### org.ojalgo.structure